+++ integration-tests/modules/__tests__/customer/store/get-me.spec.ts
@@ -11,7 +11,7 @@ const env = {}
 
 medusaIntegrationTestRunner({
   env,
-  testSuite: ({ dbConnection, getContainer, api }) => {
+  testSuite: ({ dbConnection, getContainer, api, dbUtils }) => {
     describe("GET /store/customers", () => {
       let appContainer
       let storeHeaders
@@ -20,10 +20,12 @@ medusaIntegrationTestRunner({
         appContainer = getContainer()
       })
 
-      beforeEach(async () => {
+      beforeAll(async () => {
         appContainer = getContainer()
         const publishableKey = await generatePublishableKey(appContainer)
         storeHeaders = generateStoreHeaders({ publishableKey })
+
+        await dbUtils.snapshot()
       })
 
       it("should retrieve auth user's customer", async () => {