+++ packages/admin/dashboard/src/routes/product-options/product-option-create/product-option-create.tsx
@@ -1,0 +1,10 @@
+import { RouteFocusModal } from "../../../components/modals"
+import { CreateProductOptionForm } from "./components/create-product-option-form/create-product-option-form"
+
+export const ProductOptionCreate = () => {
+  return (
+    <RouteFocusModal>
+      <CreateProductOptionForm />
+    </RouteFocusModal>
+  )
+}