+++ packages/modules/product/src/utils/index.ts
@@ -1,1 +1,3 @@
+export * from "./compute-option-link-changes"
+export * from "./compute-set-difference"
 export * from "./events"