+++ www/apps/book/sidebar.mjs
@@ -65,8 +65,8 @@ export const sidebars = [
             type: "link",
             title: "AI-Friendly Docs",
             path: "/learn/introduction/build-with-llms-ai/ai-friendly-docs",
-          }
-        ]
+          },
+        ],
       },
       {
         type: "category",
@@ -574,6 +574,11 @@ export const sidebars = [
             type: "link",
             path: "/learn/fundamentals/admin/widgets",
             title: "Admin Widgets",
+          },
+          {
+            type: "link",
+            path: "/learn/fundamentals/admin/custom-injection-zones",
+            title: "Custom Injection Zones",
           },
           {
             type: "link",